+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Nokia SRL platform driver (`Platform.NOKIA_SRL`) (#245)
+
 ---
 
 ## [3.6.0] - 2026-03-26